agent2/requirements.txt
Flask==3.1.0
requests==2.32.3
pyyaml==6.0.2
openai==1.64.0